+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -1576,6 +1576,50 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemUtilityTablesObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemUtilityTablesObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& tableId,
+    const CosemByteBuffer& length,
+    const CosemByteBuffer& buffer,
+    AttributeAccessMode mutableAccess);
+  CosemUtilityTablesObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& tableId,
+    const CosemByteBuffer& length,
+    const CosemByteBuffer& buffer,
+    AttributeAccessMode mutableAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& TableId() const;
+  const CosemByteBuffer& Length() const;
+  const CosemByteBuffer& Buffer() const;
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer tableId_;
+  CosemByteBuffer length_;
+  CosemByteBuffer buffer_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,